Fixed timeout Thread
